+++ src/__tests__/announcements/localizedAnnouncements.js
@@ -1,7 +1,7 @@
 // @flow
 import timemachine from "timemachine";
 import { localizeAnnouncements } from "../../notifications/AnnouncementProvider/logic";
-import api from "../../notifications/AnnouncementProvider/api/api.mock";
+import api from "../mocks/announements.mock";
 timemachine.config({
   dateString: "February 22, 2021 13:12:59",
 });
