+++ src/__tests__/mocks/announements.mock.js
@@ -1,7 +1,10 @@
 // @flow
 import Prando from "prando";
-import { getEnv } from "../../../env";
-import type { AnnouncementsApi, RawAnnouncement } from "../types";
+import { getEnv } from "../../env";
+import type {
+  AnnouncementsApi,
+  RawAnnouncement,
+} from "../../notifications/AnnouncementProvider/types";
 
 const announcementsPool: RawAnnouncement[] = [
   {
